requirements.txt
boto==2.48.0
Collectfast==0.5.2
Django==1.11.3
django-storages==1.6.5
django-storages-redux==1.3.3
djangorestframework==3.5.4
gunicorn==19.7.1
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pytz==2017.2
whitenoise==3.3.1
